Begin re-organizing files to follow Apollo tuts
