+++ src/resolvers.js
@@ -1,0 +1,7 @@
+module.exports = {
+  Query: {
+    pokemon: async (_source, { id }, { dataSources }) => {
+      return dataSources.pokemonAPI.getPokemon(id)
+    },
+  },
+}
