+++ src/schema.js
@@ -1,0 +1,26 @@
+const { gql } = require('apollo-server-cloudflare')
+
+module.exports = gql`
+  type PokemonSprites {
+    front_default: String!
+    front_shiny: String!
+    front_female: String!
+    front_shiny_female: String!
+    back_default: String!
+    back_shiny: String!
+    back_female: String!
+    back_shiny_female: String!
+  }
+
+  type Pokemon {
+    id: ID!
+    name: String!
+    height: Int!
+    weight: Int!
+    sprites: PokemonSprites!
+  }
+
+  type Query {
+    pokemon(id: ID!): Pokemon
+  }
+`
